Reworked BitmapProvider to be less insane. Style system revamped to allow arrays, named colours. Also removed the very silly StyleValueUber. Renamed all instances of 'colour' to 'color' cos that battle is lost. Began the RenderOp reorg towards implementing a shader cache. Introduced the notion of shader validation, i.e. when ops change render parameters they become 'invalid' have to rebind to a shader before the next draw (thus making them valid again). All automatic stuff.

+++ samples/minesweeper/app/game.cpp
@@ -31,7 +31,7 @@ void Game::restart() {
     }
 
     // Fill board with mines at random locations
-    srand((uint32_t)app.currentMillis());
+    srand(app.currentMillis());
     int mines = _rows * _cols / 7;
     _unknown = _rows * _cols - mines;
     for (int i = 0; i < mines; ++i) {
